requirements.txt
uvicorn==0.32.0
python-socketio==5.11.4
fastapi==0.115.4